# Engine/Back/public/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Engine API",
        "description": "API para gerenciar membros da equipe.",
        "contact": {
            "name": "Pablo Kauan M. Timbó",
            "email": "pablokauan.tech@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/members": {
            "get": {
                "tags": [
                    "Member"
                ],
                "operationId": "20e6baf76421bba133e9048dcdbe7355",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Member"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Member"
                ],
                "operationId": "8d4f28b9a8881fef3a26990bc41f14ad",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Member"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/members/{id}": {
            "get": {
                "tags": [
                    "Member"
                ],
                "operationId": "80f0d8d79a63de8091cbd614c69bb185",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Member"
                ],
                "operationId": "d6c64df7108dddba3753fea5e8bb06b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Member"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Member"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Member"
                ],
                "operationId": "1185c247067e00121f165f63906d5a03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Project"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Project"
                ],
                "operationId": "491d11cc6efd20335421a859159e2487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Project"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Member": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "biografia": {
                        "type": "string",
                        "example": "Biografia do membro"
                    },
                    "funcao": {
                        "type": "string",
                        "example": "Função do membro"
                    },
                    "formacao": {
                        "type": "string",
                        "example": "Formação do membro"
                    },
                    "linkedin": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "lattes": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "url_foto": {
                        "type": "string",
                        "example": "https://example.com/foto.jpg"
                    }
                },
                "type": "object"
            },
...
            "Project": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Projeto X"
                    },
                    "areas": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "subtitulo": {
                        "type": "string",
                        "example": "Breve subtítulo"
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Descrição do projeto"
                    },
                    "objetivos": {
                        "type": "string",
                        "example": "Objetivos do projeto"
                    },
                    "tecnologias": {
                        "type": "string",
                        "example": "PHP, Laravel"
                    },
                    "ano_inicio": {
                        "type": "integer",
                        "example": 2023
                    },
                    "ano_fim": {
                        "type": "integer",
                        "example": 2024
                    },
                    "financiamento": {
                        "type": "string",
                        "example": "FAPESP"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}